fix by coverity checker
Summary: Ref T6984

Test Plan: N/A

Reviewers: qinzuoyan

Reviewed By: qinzuoyan

Maniphest Tasks: T6984

Differential Revision: https://phabricator.d.xiaomi.net/D39522

+++ src/core/tools/common/simple_perf_counter_v2_fast.cpp
@@ -64,22 +64,22 @@ namespace dsn {
 
             virtual void   increment()
             {
-                uint64_t task_id = static_cast<int>(::dsn::utils::get_current_tid());
+                uint64_t task_id = static_cast<uint64_t>(::dsn::utils::get_current_tid());
                 _val[task_id % DIVIDE_CONTAINER]++;
             }
             virtual void   decrement()
             {
-                uint64_t task_id = static_cast<int>(::dsn::utils::get_current_tid());
+                uint64_t task_id = static_cast<uint64_t>(::dsn::utils::get_current_tid());
                 _val[task_id % DIVIDE_CONTAINER]--;
             }
             virtual void   add(uint64_t val)
             {
-                uint64_t task_id = static_cast<int>(::dsn::utils::get_current_tid());
+                uint64_t task_id = static_cast<uint64_t>(::dsn::utils::get_current_tid());
                 _val[task_id % DIVIDE_CONTAINER] += val;
             }
             virtual void   set(uint64_t val)
             {
-                uint64_t task_id = static_cast<int>(::dsn::utils::get_current_tid());
+                uint64_t task_id = static_cast<uint64_t>(::dsn::utils::get_current_tid());
                 _val[task_id % DIVIDE_CONTAINER] = val;
             }
             virtual double get_value()
